+++ Test/function_test.py
@@ -186,7 +186,7 @@ Example : online_help("J") or online_help(2)
 ...
 >>> online_help(4)
 ...
->>> from pycm.pycm_output import *
+>>> from pycm.output import *
 >>> from pycm.overall_funcs import *
 >>> from pycm.pycm_class_func import *
 >>> from pycm.pycm_ci import *