+++ pycm/__init__.py
@@ -2,7 +2,7 @@
 """PyCM modules."""
 from .params import PYCM_VERSION, OVERALL_BENCHMARK_LIST, CLASS_BENCHMARK_LIST
 from .pycm_error import *
-from .pycm_output import pycm_help, online_help
+from .output import pycm_help, online_help
 from .pycm_distance import DistanceType
 from .pycm_obj import ConfusionMatrix
 from .pycm_compare import Compare