+++ pycm/__main__.py
@@ -4,7 +4,7 @@
 import doctest
 import sys
 from .pycm_obj import *
-from .pycm_output import *
+from .output import *
 from art import tprint