+++ pycm/pycm_compare.py
@@ -2,7 +2,7 @@
 """Compare module."""
 from __future__ import division
 from .pycm_error import pycmCompareError
-from .pycm_output import *
+from .output import *
 from .utils import *
 from .params import *
 from .pycm_obj import ConfusionMatrix