+++ pycm/pycm_obj.py
@@ -8,7 +8,7 @@ from .pycm_handler import __imbalancement_handler__
 from .pycm_class_func import F_calc, IBA_calc, TI_calc, NB_calc, sensitivity_index_calc
 from .overall_funcs import weighted_kappa_calc, weighted_alpha_calc, alpha2_calc, brier_score_calc, log_loss_calc
 from .pycm_distance import DistanceType, DISTANCE_MAPPER
-from .pycm_output import *
+from .output import *
 from .utils import *
 from .params import *
 from .pycm_ci import __CI_overall_handler__, __CI_class_handler__