+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `pycm_profile.py` renamed to `profile.py`
 - `pycm_param.py` renamed to `params.py`
 - `pycm_overall_func.py` renamed to `overall_funcs.py`
+- `pycm_output.py` renamed to `output.py`
 ## [4.0] - 2023-06-07
 ### Added
 - `pycmMultiLabelError` class